style & fix: "代码规范化与冗余代码重构"

+++ src/pages/admin/PictureManagePage.vue
@@ -1,7 +1,7 @@
 <template>
   <div id="pictureManagePage">
     <a-flex justify="space-between">
-      <h2>图片管理</h2>
+      <h2 style="margin-bottom: 16px">图片管理</h2>
       <a-space>
         <a-button type="primary" href="/add_picture" target="_blank">+ 创建图片</a-button>
         <a-button type="primary" href="/add_picture/batch" target="_blank" ghost>+ 批量创建图片</a-button>
@@ -10,17 +10,17 @@
     <div style="margin-bottom: 16px" />
     <!-- 搜索表单 -->
     <a-form layout="inline" :model="searchParams" @finish="doSearch">
-      <a-form-item label="关键词">
+      <a-form-item label="关键词" name="searchText">
         <a-input
           v-model:value="searchParams.searchText"
           placeholder="从名称和简介搜索"
           allow-clear
         />
       </a-form-item>
-      <a-form-item label="类型">
+      <a-form-item label="类型" name="category">
         <a-input v-model:value="searchParams.category" placeholder="请输入类型" allow-clear />
       </a-form-item>
-      <a-form-item label="标签">
+      <a-form-item label="标签" name="tags">
         <a-select
           v-model:value="searchParams.tags"
           mode="tags"
@@ -29,7 +29,7 @@
           allow-clear
         />
       </a-form-item>
-      <a-form-item name="reviewStatus" label="审核状态">
+      <a-form-item label="审核状态" name="reviewStatus">
         <a-select
           v-model:value="searchParams.reviewStatus"
           style="min-width: 180px"
@@ -50,10 +50,20 @@
       :pagination="pagination"
       @change="doTableChange"
     >
+      <template #headerCell="{ column }">
+        <template v-if="column.key === 'name'">
+          <span>
+            <smile-outlined />
+            Name
+          </span>
+        </template>
+      </template>
+
       <template #bodyCell="{ column, record }">
         <template v-if="column.dataIndex === 'url'">
           <a-image :src="record.url" :width="120" />
         </template>
+        <!-- 标签 -->
         <template v-if="column.dataIndex === 'tags'">
           <a-space wrap>
             <a-tag v-for="tag in JSON.parse(record.tags || '[]')" :key="tag">
@@ -61,6 +71,7 @@
             </a-tag>
           </a-space>
         </template>
+        <!-- 图片信息 -->
         <template v-if="column.dataIndex === 'picInfo'">
           <div>格式：{{ record.picFormat }}</div>
           <div>宽度：{{ record.picWidth }}</div>
@@ -68,6 +79,7 @@
           <div>宽高比：{{ record.picScale }}</div>
           <div>大小：{{ (record.picSize / 1024).toFixed(2) }}KB</div>
         </template>
+        <!-- 审核信息 -->
         <template v-if="column.dataIndex === 'reviewMessage'">
           <div>审核状态：{{ PIC_REVIEW_STATUS_MAP[record.reviewStatus] }}</div>
           <div>审核信息：{{ record.reviewMessage }}</div>
@@ -110,6 +122,7 @@
   </div>
 </template>
 <script lang="ts" setup>
+import { SmileOutlined } from '@ant-design/icons-vue'
 import { computed, onMounted, reactive, ref } from 'vue'
 import {
   deletePictureUsingPost,
@@ -156,13 +169,8 @@ const columns = [
     dataIndex: 'picInfo',
   },
   {
-    title: '用户 id',
-    dataIndex: 'userId',
-    width: 80,
-  },
-  {
-    title: '空间 id',
-    dataIndex: 'spaceId',
+    title: '图片 id',
+    dataIndex: 'pictureId',
     width: 80,
   },
   {
@@ -217,11 +225,11 @@ onMounted(() => {
 // 分页参数
 const pagination = computed(() => {
   return {
-    current: searchParams.current,
-    pageSize: searchParams.pageSize,
+    current: searchParams.current ?? 1,
+    pageSize: searchParams.pageSize ?? 10,
     total: total.value,
     showSizeChanger: true,
-    showTotal: (total) => `共 ${total} 条`,
+    showTotal: (total: number) => `共 ${total} 条`,
   }
 })
 
@@ -234,23 +242,26 @@ const doTableChange = (page: any) => {
 
 // 搜索数据
 const doSearch = () => {
-  // 重置页码
+  // 重置搜索页码
   searchParams.current = 1
   fetchData()
 }
 
 // 删除数据
 const doDelete = async (id: string) => {
-  if (!id) {
-    return
-  }
-  const res = await deletePictureUsingPost({ id })
-  if (res.data.code === 0) {
-    message.success('删除成功')
-    // 刷新数据
-    fetchData()
-  } else {
-    message.error('删除失败')
+  // const doDelete = async (id: number) => { // 解决办法1：调整id类型; 解决办法2：调整typings.d.ts里的id类型
+  if (confirm(`确认删除id为${id}的数据吗？`)) {
+    if (!id) {
+      return
+    }
+    const res = await deletePictureUsingPost({ id }) // id报红是因为这里定义的string与typings.d.ts里的number不匹配
+    if (res.data.code === 0) {
+      message.success('删除成功')
+      // 刷新数据
+      fetchData()
+    } else {
+      message.error('删除失败')
+    }
   }
 }
